# tochar: implement cache for parse format