--- a/pkg/sql/distsql/server.go
+++ b/pkg/sql/distsql/server.go
@@ -40,6 +40,7 @@ import (
 	"github.com/cockroachdb/cockroach/pkg/util/mon"
 	"github.com/cockroachdb/cockroach/pkg/util/pprofutil"
 	"github.com/cockroachdb/cockroach/pkg/util/timeutil"
+	"github.com/cockroachdb/cockroach/pkg/util/tochar"
 	"github.com/cockroachdb/cockroach/pkg/util/tracing"
 	"github.com/cockroachdb/cockroach/pkg/util/tracing/grpcinterceptor"
 	"github.com/cockroachdb/errors"
@@ -63,10 +64,11 @@ var noteworthyMemoryUsageBytes = envutil.EnvOrDefaultInt64("COCKROACH_NOTEWORTHY
 // ServerImpl implements the server for the distributed SQL APIs.
 type ServerImpl struct {
 	execinfra.ServerConfig
-	flowRegistry     *flowinfra.FlowRegistry
-	remoteFlowRunner *flowinfra.RemoteFlowRunner
-	memMonitor       *mon.BytesMonitor
-	regexpCache      *tree.RegexpCache
+	flowRegistry      *flowinfra.FlowRegistry
+	remoteFlowRunner  *flowinfra.RemoteFlowRunner
+	memMonitor        *mon.BytesMonitor
+	regexpCache       *tree.RegexpCache
+	toCharFormatCache *tochar.FormatCache
 }
 
 var _ execinfrapb.DistSQLServer = &ServerImpl{}
@@ -76,10 +78,11 @@ func NewServer(
 	ctx context.Context, cfg execinfra.ServerConfig, remoteFlowRunner *flowinfra.RemoteFlowRunner,
 ) *ServerImpl {
 	ds := &ServerImpl{
-		ServerConfig:     cfg,
-		regexpCache:      tree.NewRegexpCache(512),
-		flowRegistry:     flowinfra.NewFlowRegistry(),
-		remoteFlowRunner: remoteFlowRunner,
+		ServerConfig:      cfg,
+		regexpCache:       tree.NewRegexpCache(512),
+		toCharFormatCache: tochar.NewFormatCache(512),
+		flowRegistry:      flowinfra.NewFlowRegistry(),
+		remoteFlowRunner:  remoteFlowRunner,
 		memMonitor: mon.NewMonitor(
 			"distsql",
 			mon.MemoryResource,
@@ -354,6 +357,7 @@ func (ds *ServerImpl) setupFlow(
 			NodeID:                    ds.ServerConfig.NodeID,
 			Codec:                     ds.ServerConfig.Codec,
 			ReCache:                   ds.regexpCache,
+			ToCharFormatCache:         ds.toCharFormatCache,
 			Locality:                  ds.ServerConfig.Locality,
 			Tracer:                    ds.ServerConfig.Tracer,
 			Planner:                   &faketreeeval.DummyEvalPlanner{Monitor: monitor},