@otan 2022-12-09
made it RWMutex

bors r=rafiss